llm/requirements.txt
bottleneck==1.3.7
brotlipy==0.7.0
faiss-cpu==1.7.4
flake8==6.0.0
langchain==0.0.244
openai==0.27.8
pyopenssl==23.2.0
pysocks==1.7.1
python-dotenv==1.0.0
streamlit==1.25.0
tiktoken==0.4.0
unstructured==0.8.4
